### 2021-09-23 #3758 @whedon
[PRE REVIEW]: AIBECS.jl: The ideal tool for exploring global marine biogeochemical cycles.
**Submitting author:** @briochemc (<a href="http://orcid.org/0000-0002-3838-5976">Benoit Pasquier</a>)
**Repository:** <a href="https://github.com/JuliaOcean/AIBECS.jl" target ="_blank">https://github.com/JuliaOcean/AIBECS.jl</a>
**Version:** v0.10.10
