### 2021-09-23 #3758 @kthyng
Hi @briochemc! Please check out the [reviewer list](https://bit.ly/joss-reviewers) and think through your own network of people you know (who aren't conflicted) to let me know some suggestions for reviewers. List their github handles here without the "@" prefix so they aren't preemptively pinged.
