### 2021-10-11 #3758 @kthyng
Hi @gregbritten! Are you interested in reviewing this submission for the Journal of Open Source Software? The reviews are done openly in github via issues in the software and a summary in a review issue. You can read more about the process [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html). We are asking for reviews to be finished in 4-6 weeks, but suggest you start sooner than that since the review process tends to be iterative between the reviewer and author. Thanks for your consideration.
