### 2021-10-11 #3758 @dankelley
Yes, I can review it.  (I've already started and will likely finish this afternoon.  It's a holiday here, and I have a break from term but only for a day.)
