### 2021-10-11 #3814 @whedon
[REVIEW]: AIBECS.jl: The ideal tool for exploring global marine biogeochemical cycles.
**Submitting author:** @briochemc (<a href="http://orcid.org/0000-0002-3838-5976">Benoit Pasquier</a>)
**Repository:** <a href="https://github.com/JuliaOcean/AIBECS.jl" target ="_blank">https://github.com/JuliaOcean/AIBECS.jl</a>
**Version:** v0.11.2

@dankelley, @zhenwu0728 please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
