### 2021-10-14 1fd5795
Fix transportoperator docstring
