### 2021-11-02 #3758 @patrickcgray
> @patrickcgray would you be able to review this JOSS submission? The reviews are done openly in github via issues in the software and a summary in a review issue. You can read more about the process [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html). We are asking for reviews to be finished in 4-6 weeks, but suggest you start sooner than that since the review process tends to be iterative between the reviewer and author. Thanks for your consideration.

Hi @kthyng this package looks amazing! and I'd love to dig into it, but I'm not familiar enough with Julia and don't have the time over the coming month to do it justice. I look forward to seeing it in JOSS though!
