### 2021-11-02 #3758 @kthyng
Thanks for your response @patrickcgray!

@zhenwu0728 I wanted to come back to you. Regarding the workshop you ran with @briochemc — do you feel like you could be an unbiased reviewer of the software package? If so, I think we could proceed given that we are airing the conflict of interest in the open.
