### 2021-11-08 JuliaOcean/AIBECS.jl#83 @zhenwu0728
Comments on JOSSpaper/paper.md
Comments for openjournals/joss-reviews#3814

I've worked through "ideal age" and "PO4-POP" examples, and they are well documented and informative.
I only have a few minor comments for the package and the paper.

- The examples are great, but Binder is not working properly (tried a few times to launch Binder and it crashed while running the model, possibly because of the limit of memory). I suggest the authors may use Pluto.jl and [PlutoSliderServer.jl](https://github.com/JuliaPluto/PlutoSliderServer.jl) to host examples.
- A better `show()` is needed for `OCIM2.load()`, `F_and_∇ₓF`, etc. Now it's hard to get any useful information from the output.

For paper.md:
- L11-15: It might be good to provide one or two examples for both simple and complicated models.
- L16: It would be good to point out that AIBECS.jl only resolves steady states.
- L31: It is worth mentioning that minimizing model-observation mismatches may cause overfitting.
- L73: `Taburet_etal_2019` citation is broken.
- L102-119: I think there are too many citations, only one or two for each kind of model is enough.
