### 2021-11-19 briochemc/AIBECS.jl#83 @briochemc
I gave a bit of thought to the issue with `F_and_∇ₓF` and think I have a decent solution. There is the `AIBECSFunction` that one can use instead, which returns a `SciMLBase.ODEFunction`. I'm writing an update on which one would do

```julia
julia> F = AIBECSFunction(T, G)
(::SciMLBase.ODEFunction{false, AIBECS.var"#f#438"{Tuple{AIBECS.var"#51#52"{SparseMatrixCSC{Float64, Int64}}}, Vector{Int64}, AIBECS.var"#G#436"{Tuple{typeof(G)}, AIBECS.var"#tracers#434"{Int64, Int64}}, AIBECS.var"#tracer#435"{Int64, Int64}}, LinearAlgebra.UniformScaling{Bool}, Nothing, Nothing, AIBECS.var"#jac#443"{AIBECS.var"#T#440"{Tuple{AIBECS.var"#51#52"{SparseMatrixCSC{Float64, Int64}}}, Int64, Vector{Int64}}, AIBECS.var"#∇ₓG#439"{Tuple{typeof(G)}, Int64, Int64}}, Nothing, Nothing, Nothing, Nothing, Nothing, Nothing, Nothing, Nothing, Nothing, typeof(SciMLBase.DEFAULT_OBSERVED), Nothing}) (generic function with 7 methods)
```

and then this `F` can be called directly to get the tendencies:

```julia
julia> F(x,p)
200160-element Vector{Float64}:
 1.0
 1.0
 ⋮
 1.0
 1.0
```

The Jacobian can be accessed using the SciML syntax:

```julia
julia> F(Val{:jac},x,p)
200160×200160 SparseMatrixCSC{Float64, Int64} with 3018260 stored entries:
⢿⣷⣄⠀⠀⠀⠀⠀⠀⠀
⠀⠙⢿⣷⣄⠀⠀⠀⠀⠀
⠀⠀⠀⠙⢿⣷⣄⠀⠀⠀
⠀⠀⠀⠀⠀⠙⢿⣷⣄⠀
⠀⠀⠀⠀⠀⠀⠀⠙⢿⣷
```

But I don't need to have any discussion about the Jacobian and the solvers in the tutorials. So in that sense, having just `F = AIBECSFunction(T, G)` and letting the user call `F(x,p)` seems friendlier. (Of course AIBECS can and will access the Jacobian under the hood to solve the problem without bothering the user with the implementation details.)

The steady-state problem is then set up with only `F` (which is also simpler, no need to write `∇ₓF` at all from a user's point of view:

```julia
julia> prob = SteadyStateProblem(F, x, p)
SteadyStateProblem with uType Vector{Float64}. In-place: false
u0: 200160-element Vector{Float64}:
 0.0
 0.0
 ⋮
 0.0
 0.0
```

Would that be better in your opinion?
