+++ src/multiTracer.jl
@@ -28,10 +28,6 @@ function Base.:*(A::LinearOperators, u::AbstractArray) # has to be same type!
 end
 export LinearOperators
 
-# What would be the usage like
-# FUN = AIBECSfunction(...)
-# F = FUN.f
-# ∇ₓF = FUN.jac
 # AIBECSFunction creates an ODEFunction from the Ts and Gs
 AIBECSFunction(T::AbstractSparseArray, G::Function) = AIBECSFunction(p -> T, G, T.n)
 AIBECSFunction(T::Function, G::Function, nb::Int) = AIBECSFunction(T, (G,), nb)
@@ -57,7 +53,7 @@ function iipAIBECSFunction(Ts, Gs, nb, nt=length(Ts), Tidx=1:nt)
         end
         du
     end
-    function f(du, u, p, t)
+    function f(du, u, p, t=0)
         G(du, u, p)
         for jT in eachindex(Ts)
             op = Ts[jT](p)
@@ -67,21 +63,20 @@ function iipAIBECSFunction(Ts, Gs, nb, nt=length(Ts), Tidx=1:nt)
         end
         du
     end
-    f(u, p, t) = (du = copy(u); f(du, u, p, t); du)
     # Jacobian
     ∇ₓG(u, p) = inplace_local_jacobian(Gs, u, p, nt, nb)
     function T(p)
         uniqueTs = [Tⱼ(p) for Tⱼ in Ts]
         blockdiag([uniqueTs[Tidx[j]] for j in 1:nt]...)
     end
-    jac(u, p, t) = ∇ₓG(u, p) - T(p)
+    jac(u, p, t=0) = ∇ₓG(u, p) - T(p)
     return ODEFunction{true}(f, jac=jac)
 end
 function oopAIBECSFunction(Ts, Gs, nb, nt=length(Ts), Tidx=1:nt)
     tracers(u) = state_to_tracers(u, nb, nt)
     tracer(u, i) = state_to_tracer(u, nb, nt, i)
     G(u, p) = reduce(vcat, Gⱼ(tracers(u)..., p) for Gⱼ in Gs)
-    function f(u, p, t)
+    function f(u, p, t=0)
         du = copy(G(u, p))
         for jT in eachindex(Ts)
             op = Ts[jT](p)
@@ -97,50 +92,54 @@ function oopAIBECSFunction(Ts, Gs, nb, nt=length(Ts), Tidx=1:nt)
         uniqueTs = [Tⱼ(p) for Tⱼ in Ts]
         blockdiag([uniqueTs[Tidx[j]] for j in 1:nt]...)
     end
-    jac(u, p, t) = ∇ₓG(u, p) - T(p)
+    jac(u, p, t=0) = ∇ₓG(u, p) - T(p)
     return ODEFunction(f, jac=jac)
 end
-# This AIBECSFunction overloads F and ∇ₓF for λ instead of p
+# AIBECSFunction calls itself to allow for both λ::Vector and p::APar
 function AIBECSFunction(Ts, Gs, nb::Int, ::Type{P}) where {P <: APar}
     AIBECSFunction(AIBECSFunction(Ts, Gs, nb), P)
 end
 function AIBECSFunction(fun::ODEFunction{false}, ::Type{P}) where {P <: APar}
-    jac(u, p::P, t) = fun.jac(u, p, t)
-    jac(u, λ::Vector, t) = fun.jac(u, λ2p(P, λ), t)
-    f(u, p::P, t) = fun.f(u, p, t)
-    f(u, λ::Vector, t) = fun.f(u, λ2p(P, λ), t)
[…]
