+++ previews/PR85/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"tutorials/5_river_discharge/","page":"River discharge","title":"River discharge","text":"EditURL = \"https://github.com/JuliaOcean/AIBECS.jl/blob/master/docs/lit/tutorials/5_river_discharge.jl\"","category":"page"},{"location":"tutorials/5_river_discharge/#river-discharge","page":"River discharge","title":"River discharge","text":"","category":"section"},{"location":"tutorials/5_river_discharge/","page":"River discharge","title":"River discharge","text":"(Image: ) (Image: )","category":"page"},{"location":"tutorials/5_river_discharge/","page":"River discharge","title":"River discharge","text":"In this tutorial we will simulate a fictitious radioactive tracer that is injected into the ocean by the 200 largest rivers (by estimated discharge). The 200 major rivers dataset from Dai and Trenberth (2002) is available from within the AIBECS. Once \"born\", our ficitious tracer decays with a parameter timescale tau as it flows through ocean basins.","category":"page"},{"location":"tutorials/5_river_discharge/","page":"River discharge","title":"River discharge","text":"The 3D tracer equation is:","category":"page"},{"location":"tutorials/5_river_discharge/","page":"River discharge","title":"River discharge","text":"leftfracpartialpartial t + nabla cdot (boldsymbolu + mathbfKnabla)right x = s_mathsfrivers - x  tau","category":"page"},{"location":"tutorials/5_river_discharge/","page":"River discharge","title":"River discharge","text":"where nabla cdot left boldsymbolu - mathbfK nabla right represents the ocean circulation transport. (Tracer transport operators are described in the documentation.) The riverine source of the tracer is s_mathsfrivers, and x  tau is the decay rate.","category":"page"},{"location":"tutorials/5_river_discharge/","page":"River discharge","title":"River discharge","text":"In AIBECS, we must recast this equation in the generic form","category":"page"},{"location":"tutorials/5_river_discharge/","page":"River discharge","title":"River discharge","text":"leftfracpartialpartial t + mathbfT(boldsymbolp)right boldsymbolx = boldsymbolG(boldsymbolxboldsymbolp)","category":"page"},{"location":"tutorials/5_river_discharge/","page":"River discharge","title":"River discharge","text":"We start by telling Julia that we want to use the AIBECS and the OCIM2 transport matrix for the ocean circulation.","category":"page"},{"location":"tutorials/5_river_discharge/","page":"River discharge","title":"River discharge","text":"using AIBECS\ngrd, T_OCIM2 = OCIM2.load()","category":"page"},{"location":"tutorials/5_river_discharge/","page":"River discharge","title":"River discharge","text":"For the radioactive decay, we simply use","category":"page"},{"location":"tutorials/5_river_discharge/","page":"River discharge","title":"River discharge","text":"function decay(x, p)\n    @unpack τ = p\n    return x / τ\nend","category":"page"},{"location":"tutorials/5_river_discharge/","page":"River discharge","title":"River discharge","text":"To build the river sources, we will load the geographic locations and discharge (in m³ s⁻¹) from the Dai and Trenberth (2017) dataset.","category":"page"},{"location":"tutorials/5_river_discharge/","page":"River discharge","title":"River discharge","text":"RIVERS = Rivers.load()","category":"page"},{"location":"tutorials/5_river_discharge/","page":"River discharge","title":"River discharge","text":"This is an array of rivers, for which the type River{T} contains the river's name, lat–lon coordinates, and discharge in m³ s⁻¹. For example, the first river 
[…]
