### 2021-11-19 briochemc/AIBECS.jl#83 @briochemc
@zhenwu0728 FYI, there is now [an ideal-age tutorial deployed from the linked PR](https://juliaocean.github.io/AIBECS.jl/previews/PR85/tutorials/1_ideal_age/) which shows how this looks when using `AIBECSFunction` instead of `F_and_∇ₓF`. Does that seem like a good solution for the output of `F_and_∇ₓF`?
