### 2021-11-19 briochemc/AIBECS.jl#83 @zhenwu0728
Sorry @briochemc I didn't make it clear enough.
I realized that the outputs(OceanGrid, SparseMatrix etc.) are not the type you defined in AIBECS.jl so you don't have control.

The `AIBECSFunction` is absolutely better.

But `grd, TOCIM2 = OCIM2.load()` still doesn't show `OceanGrid` information in the example from your last comment.

> Thus maybe the issue for OCIM2.load() is actually the extra outputs (which are not always assigned to a variable)? These include He fluxes as unitful vectors (as explained in the docstring that I pasted above). Below I assign all outputs from OCIM2.load() and I am guessing that issue with the output of OCIM2.load() might be these two unitful vectors which get printed inline inside the tuple of outputs? (See the very long line of output below)

Yes, if the extra outputs are not used, could they be removed from the `return`?
