### 2021-11-22 f3fb682
Fix convoluted/messy output from `F_and_∇ₓF` and `load` (#85)
* JOSS submission WIP

* Move from F_and_∇ₓF to AIBECSFunction API

* Remove default He output from OCIM2.load
