+++ docs/paper.md
@@ -1,0 +1,167 @@
+---
+title: 'AIBECS.jl: The ideal tool for exploring global marine biogeochemical cycles.'
+tags:
+  - Julia
+  - Oceanography
+  - Biogeochemical Cycles
+  - Earth Sciences
+authors:
+  - name: Benoît Pasquier^[corresponding author]
+    orcid: 0000-0002-3838-5976
+    affiliation: "1" # (Multiple affiliations must be quoted)
+  - name: Seth G. John
+    orcid: 0000-0002-8257-626X
+    affiliation: "1" # (Multiple affiliations must be quoted)
+  - name: François W. Primeau
+    orcid: 0000-0001-7452-9415
+    affiliation: "2"
+affiliations:
+  - name: Department of Earth Sciences, University of Southern California
+    index: 1
+  - name: Department of Earth System Science, University of California, Irvine
+    index: 2
+date: 19 Aug 2021
+bibliography: paper.bib
+
+---
+
+# Summary
+
+
+
+
+The ocean transports, mixes, and transforms chemical constituents on a multitude of time and length scales.
+Observations and models are both essential for making sense of this complex system.
+However, the days of just publishing data without any quantitative modelling are over, increasing pressure on sea-going oceanographers that are expected to be proficient in the use of biogeochemical models.
+And while the simplest of these models consist of only a few boxes, with solutions that are easily obtained by hand or on a simple desktop computer, the most advanced models have high-resolution three-dimensional meshes that require high-performance computing (HPC) clusters and a considerable amount of computational-science expertise.
+A high barrier to entry for modelling in oceanography hinders advances in the field.
+[AIBECS.jl](https://github.com/JuliaOcean/AIBECS.jl) [@Pasquier:2021] is a [JuliaOcean](https://github.com/JuliaOcean/)-affiliated package written in [Julia](https://julialang.org/) [@Bezanson_etal:2017], which aims to lower this barrier by providing an easy-to-use, open-source, and modular framework for simulating global marine tracers and biogeochemical parameter fitting.
+
+
+
+
+
+
+A conceptual model of the cycle of any marine tracer essentially requires two components.
+(i) A model of how the tracer is transported (be it the ocean currents and eddies, gravitational settling, or a combination of those), and (ii) a model of the local sources and sinks at any location.
+AIBECS.jl is built on this concept and allows users to build numerical models of marine tracers by selecting a circulation and/or vertical transport of the tracer with particles and local sources and sinks.
+Tools for generating the equations, solving them, and then diagnosing and plotting the simulated tracers are also provided, either directly by AIBECS.jl, by its dependencies, or by satellite packages in the AIBECS and Julia ecosystem.
+
+
+
+
+By definition, all models are wrong because they make simplifying assumptions [@Box_1979].
+These assumptions give rise to model parameters that can capture fundamental characteristics of the system.
+Once the model of the cycle of a marine tracer is implemented, it is thus natural to try to estimate biogeochemical parameters by minimizing model–observation mismatches, which also improves the skill of the model.
+Parameter optimization is generally more efficiently performed when first and second derivatives are available, and these derivatives can also be used to estimate parameter sensitivity [@Thacker_1989].
+AIBECS.jl was designed with parameter fitting in mind and provides, along with satellite packages, tools for generating first- and second-order derivatives automatically and efficiently.
+
+
+
[…]
