+++ docs/paper.bib
@@ -1,0 +1,996 @@
+@article{Amante_Eakins_2009,
+  title     = {{ETOPO1} 1 arc-minute global relief model: procedures, data sources and analysis},
+  author    = {Amante, Christopher and Eakins, Barry W},
+  year      = {2009},
+  journal   = {NOAA Technical Memorandum NESDIS NGDC-24},
+  publisher = {National Geophysical Data Center, NOAA},
+  doi       = {10.7289/V5C8276M}
+}
+
+
+@article{Archer_etal:2000,
+  author  = {Archer, David E. and Eshel, Gidon and Winguth, Arne and Broecker, Wallace and Pierrehumbert, Ray and Tobis, Michael and Jacob, Robert},
+  title   = {Atmospheric p{CO}\textsubscript{2} sensitivity to the biological pump in the ocean},
+  journal = {Global Biogeochemical Cycles},
+  volume  = {14},
+  number  = {4},
+  pages   = {1219--1230},
+  doi     = {10.1029/1999GB001216},
+  year    = {2000}
+}
+
+@article{Bardin_etal_OM_2014,
+  title    = {An offline implicit solver for simulating prebomb radiocarbon},
+  journal  = {Ocean Modelling},
+  volume   = {73},
+  pages    = {45-58},
+  year     = {2014},
+  issn     = {1463-5003},
+  doi      = {10.1016/j.ocemod.2013.09.008},
+  author   = {Ann Bardin and François Primeau and Keith Lindsay},
+  keywords = {Implicit solver, Prebomb radiocarbon, Newton–Krylov, Preconditioner, Global ocean modeling}
+}
+
+@article{Bardin_etal_OM_2016,
+  title    = {Evaluation of the accuracy of an offline seasonally-varying matrix transport model for simulating ideal age},
+  journal  = {Ocean Modelling},
+  volume   = {105},
+  pages    = {25-33},
+  year     = {2016},
+  issn     = {1463-5003},
+  doi      = {10.1016/j.ocemod.2016.07.003},
+  author   = {Ann Bardin and François Primeau and Keith Lindsay and Andrew Bradley},
+  keywords = {Implicit solver, Ideal age, Newton–Krylov, Offline, Transport matrix, Global ocean modeling}
+}
+
+@article{Besancon_etal_2021,
+  author   = {Mathieu Besançon and Theodore Papamarkou and David Anthoff and Alex Arslan and Simon Byrne and Dahua Lin and John Pearson},
+  title    = {Distributions.jl: Definition and Modeling of Probability Distributions in the JuliaStats Ecosystem},
+  journal  = {Journal of Statistical Software},
+  volume   = {98},
+  number   = {16},
+  year     = {2021},
+  keywords = {Julia; distributions; modeling; interface; mixture; KDE; sampling; probabilistic programming; inference},
+  issn     = {1548-7660},
+  pages    = {1--30},
+  doi      = {10.18637/jss.v098.i16}
+}
+
+@article{Bezanson_etal:2017,
+  author  = {Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B.},
+  title   = {Julia: A Fresh Approach to Numerical Computing},
+  journal = {SIAM Review},
+  volume  = {59},
+  number  = {1},
+  pages   = {65-98},
+  year    = {2017},
+  doi     = {10.1137/141000671}
+}
+
+@incollection{Box_1979,
+  title     = {Robustness in the Strategy of Scientific Model Building},
+  editor    = {Robert L. Launer and Graham N. Wilkinson},
+  booktitle = {Robustness in Statistics},
+  publisher = {Academic Press},
+  pages     = {201-236},
+  year      = {1979},
+  isbn      = {978-0-12-438150-6},
[…]
