### 2021-11-22 #3814 @briochemc
@kthyng Yes! I've been working (slowly) on the code revisions according to @zhenwu0728's code comments! (I already merged a PR with one fix for output display, and I am trying options for moving the notebooks from Jupyter to Pluto)
