+++ docs/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'AIBECS.jl: The ideal tool for exploring global marine biogeochemical cycles.'
+title: 'AIBECS.jl: A tool for exploring global marine biogeochemical cycles.'
 tags:
   - Julia
   - Oceanography
@@ -33,9 +33,9 @@ bibliography: paper.bib
 The ocean transports, mixes, and transforms chemical constituents on a multitude of time and length scales.
 Observations and models are both essential for making sense of this complex system.
 However, the days of just publishing data without any quantitative modelling are over, increasing pressure on sea-going oceanographers that are expected to be proficient in the use of biogeochemical models.
-And while the simplest of these models consist of only a few boxes, with solutions that are easily obtained by hand or on a simple desktop computer, the most advanced models have high-resolution three-dimensional meshes that require high-performance computing (HPC) clusters and a considerable amount of computational-science expertise.
+And while the simplest of these models consist of only a few boxes [e.g., the two-box model of @Archer_etal_GBC_2000], with solutions that are easily obtained by hand or on a simple desktop computer, the most advanced models have high-resolution three-dimensional meshes that require high-performance computing (HPC) clusters and a considerable amount of computational-science expertise [e.g., the MITgcm, @Campin_etal_2021].
 A high barrier to entry for modelling in oceanography hinders advances in the field.
-[AIBECS.jl](https://github.com/JuliaOcean/AIBECS.jl) [@Pasquier:2021] is a [JuliaOcean](https://github.com/JuliaOcean/)-affiliated package written in [Julia](https://julialang.org/) [@Bezanson_etal:2017], which aims to lower this barrier by providing an easy-to-use, open-source, and modular framework for simulating global marine tracers and biogeochemical parameter fitting.
+[AIBECS.jl](https://github.com/JuliaOcean/AIBECS.jl) [@Pasquier:2021] is a [JuliaOcean](https://github.com/JuliaOcean/)-affiliated package written in [Julia](https://julialang.org/) [@Bezanson_etal:2017], which aims to lower this barrier by providing an easy-to-use, open-source, and modular framework for simulating global marine tracers in steady-state and biogeochemical parameter fitting.
 
 
 
@@ -45,21 +45,20 @@ A high barrier to entry for modelling in oceanography hinders advances in the fi
 A conceptual model of the cycle of any marine tracer essentially requires two components.
 (i) A model of how the tracer is transported (be it the ocean currents and eddies, gravitational settling, or a combination of those), and (ii) a model of the local sources and sinks at any location.
 AIBECS.jl is built on this concept and allows users to build numerical models of marine tracers by selecting a circulation and/or vertical transport of the tracer with particles and local sources and sinks.
-Tools for generating the equations, solving them, and then diagnosing and plotting the simulated tracers are also provided, either directly by AIBECS.jl, by its dependencies, or by satellite packages in the AIBECS and Julia ecosystem.
+Tools for generating the steady-state equations, solving them, and then diagnosing and plotting the simulated tracers are also provided, either directly by AIBECS.jl, by its dependencies, or by satellite packages in the AIBECS and Julia ecosystem.
 
 
 
-
-By definition, all models are wrong because they make simplifying assumptions [@Box_1979].
-These assumptions give rise to model parameters that can capture fundamental characteristics of the system.
[…]
