+++ docs/lit/tutorials/2_radiocarbon.jl
@@ -6,6 +6,11 @@
 #md # [![](https://mybinder.org/badge_logo.svg)](@__BINDER_ROOT_URL__/notebooks/2_radiocarbon.ipynb)
 #md # [![](https://img.shields.io/badge/show-nbviewer-579ACA.svg)](@__NBVIEWER_ROOT_URL__/notebooks/2_radiocarbon.ipynb)
 
+#md # !!! note
+#md #     All the AIBECS tutorials and how-to guides are available as Jupyter notebooks.
+#md #     You can view them with [nbviewer](https://nbviewer.jupyter.org/)or execute them online with [binder](https://mybinder.org/) by clicking on the badges above!
+#md #     (Note that binder can be slow to launch and its memory caps can be a problem when running.)
+
 # In this tutorial, we will simulate the radiocarbon age using the AIBECS by
 # 1. defining the transport `T(p)` and the sources and sinks `G(x,p)`,
 # 1. defining the parameters `p`,
