+++ src/overload_solve.jl
@@ -1,9 +1,8 @@
-using DiffEqBase
 
-struct CTKAlg <: DiffEqBase.AbstractSteadyStateAlgorithm end
+struct CTKAlg <: SciMLBase.AbstractSteadyStateAlgorithm end
 
 """
-    solve(prob::DiffEqBase.AbstractSteadyStateProblem,
+    solve(prob::SciMLBase.AbstractSteadyStateProblem,
           alg::CTKAlg;
           nrm=norm,
           τstop=1e12*365*24*60*60,
@@ -12,18 +11,18 @@ struct CTKAlg <: DiffEqBase.AbstractSteadyStateAlgorithm end
 
 Solves `prob` using the modified C.T.Kelley Shamanskii algorithm.
 """
-function DiffEqBase.solve(prob::DiffEqBase.AbstractSteadyStateProblem,
+function SciMLBase.solve(prob::SciMLBase.AbstractSteadyStateProblem,
                           alg::CTKAlg;
                           nrm=norm,
                           τstop=ustrip(u"s", 1e6u"Myr"),
                           preprint="",
                           maxItNewton=50)
-    # Define the functions according to DiffEqBase.SteadyStateProblem type
+    # Define the functions according to SciMLBase.SteadyStateProblem type
     p = prob.p
     x0 = copy(prob.u0)
     dx = copy(x0)
     function F(x)
-        if DiffEqBase.isinplace(prob)
+        if SciMLBase.isinplace(prob)
             prob.f(dx, x, p)
             dx
         else
@@ -31,11 +30,11 @@ function DiffEqBase.solve(prob::DiffEqBase.AbstractSteadyStateProblem,
         end
     end
     ∇ₓF(x) = prob.f.jac(x, p)
-    # Compute `u_steady` and `resid` as per DiffEqBase using my algorithm
+    # Compute `u_steady` and `resid` as per SciMLBase using my algorithm
     x_steady = NewtonChordShamanskii(F, ∇ₓF, nrm, x0, τstop; preprint=preprint, maxItNewton=maxItNewton)
     resid = F(x_steady)
-    # Return the common DiffEqBase solution type
-    DiffEqBase.build_solution(prob, alg, x_steady, resid; retcode=:Success)
+    # Return the common SciMLBase solution type
+    SciMLBase.build_solution(prob, alg, x_steady, resid; retcode=:Success)
 end
 
 export solve, SteadyStateProblem, CTKAlg
