+++ src/AIBECS.jl
@@ -27,6 +27,8 @@ using NearestNeighbors
 using Shapefile
 using Bijectors
 using ImageFiltering
+using SciMLBase
+
 
 
 
