+++ Project.toml
@@ -9,7 +9,6 @@ CodecZlib = "944b1d66-785c-5afd-91f1-9de20f533193"
 DataDeps = "124859b0-ceae-595e-8997-d05f6a7a8dfe"
 DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
 Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
-DiffEqBase = "2b5f629d-d688-5b77-993f-72d75c75574e"
 Distances = "b4f34e82-e78d-54a5-968a-f98e89d6e8f7"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 Downloads = "f43a241f-c20a-4ad4-852c-f6b1247861c6"
@@ -27,6 +26,7 @@ OceanGrids = "cfe838f4-859f-11e9-2ea1-df7d4e7c3537"
 Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
 RecipesBase = "3cdcf5f2-1ef4-517c-9805-6587b60abb01"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
+SciMLBase = "0bca4576-84f4-4d90-8ffe-ffa030f20462"
 Shapefile = "8e980c4a-a4fe-5da2-b3a7-4b4b0353a2f4"
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 StringDistances = "88034a9c-02f8-509d-84a9-84ec65e18404"
@@ -40,7 +40,6 @@ Bijectors = "0.8, 0.9"
 CodecZlib = "0.7"
 DataDeps = "0.7"
 DataFrames = "0.20, 0.21, 0.22, 1.0"
-DiffEqBase = "6"
 Distances = "0.9, 0.10"
 Distributions = "0.23, 0.24, 0.25"
 FieldMetadata = "0.3"
@@ -55,6 +54,7 @@ NearestNeighbors = "0.4"
 OceanGrids = "0.4"
 RecipesBase = "1"
 Reexport = "0.2, 1.0"
+SciMLBase = "1"
 Shapefile = "0.6, 0.7"
 StringDistances = "0.10, 0.11"
 UnPack = "1"
