### 2021-12-10 #3814 @briochemc
Dear @kthyng,

We have revised the manuscript according to @dankelley's and @zhenwu0728 's review comments in [PR #87](https://github.com/JuliaOcean/AIBECS.jl/pull/87) (for convenience: [link to the diff of the manuscript since its first submission](https://github.com/JuliaOcean/AIBECS.jl/compare/04740b960819a74f7899ab5b26f7596fe0b009df...JOSSpaper#diff-b78ae3bcb0e076c47930040ffb0e3bd3cb6a3aaa232b98a678557bbebfe71e4d)).

We have fixed all the code issues mentioned in [Issue #83](https://github.com/JuliaOcean/AIBECS.jl/issues/83)  by @zhenwu0728 in [PR #85](https://github.com/JuliaOcean/AIBECS.jl/pull/85) and [PR #87](https://github.com/JuliaOcean/AIBECS.jl/pull/87) except for the suggestion to port the notebooks from Jupyter to Pluto. This is because of breaking changes in Pluto itself that made it incompatible with the current parameter-handling implementation of AIBECS, which relies on function-overloading macros. While we hope to eventually revisit the parameter internals and interface to rely on fewer macros and make it Pluto compatible, we believe such extensive code revisions are out of the scope of @zhenwu0728's suggestion. We, therefore, hope that our revisions in [PR #85](https://github.com/JuliaOcean/AIBECS.jl/pull/85) are sufficient. (It makes sense for [Issue #83](https://github.com/JuliaOcean/AIBECS.jl/issues/83) to remain open, however, as a reminder to make the transition from Jupyter to Pluto in the future).

Overall, we believe @dankelley's and @zhenwu0728 's comments have helped us improve the software and the manuscript. (They are now acknowledged in the manuscript.)

Below you will find our point-by-point replies to the reviewers' comments.

We hope you will find that our revisions are sufficient for publication in JOSS.

On behalf of the authors,
Best Regards,
Benoit Pasquier

## Response to reviewer 1 (@dankelley) comments

> I recommend accepting this submission. It is a welcome addition to the field, and it would be to the credit of JOSS to publicize it to a wide range of educators/researchers.

We thank @dankelley for his recommendation.

> I have made some comments below, in the major/minor format that I normally use for scientific journals. I am not sure whether this is useful for a JOSS manuscript, but I didn't see any other way to provide more detailed comments, after I finished the checklist. As can be seen, I have a few suggestions that might improve the clarity of the writing, but I leave it up to the authors to decide whether to follow them (except for the minor comment about a broken citation).

The broken citation has been fixed and the suggestions have been broadly followed (details below).

> Were this a conventional scientific paper, I would have appreciated a bit more about the class of modelling frameworks supported by this project, and about any related limitations. However, for JOSS, I think what the authors have said on this matter is wholly sufficient.

We have added "steady-state" in multiple sentences to clarify (repeatedly) that this is the class of modelling currently in the scope of AIBECS.

> As for JOSS, I do think it would be helpful to add a few sentences about why Julia is chosen for the project, because their addition might give the work a broader impact. (Basically, Julia opens up doors that are closed to other systems, and folks who have not used it might not be aware just how important this is.) However, I leave the choice up to the authors.

We have added a sentence clarifying that AIBECS owes its user-friendliness and efficiency to the Julia language.

> I am always happy to see software aimed at more than just a single project. It's even better when developers take the time to educate others about how their creations might help others in the work. I think JOSS is designed specifically for this sort of thing, and I think the authors deserved to be thanked not just for their authorship of the software, but also for considering JOSS as a way to expose the work to a wider audience.

We thank @dankelley for this comment.

> ### Major Comments
>
> The manuscript outlines in reasonable detail the need for the package, although I would have appreciated more commentary on the focus on steady-state models, which is a limitation that is certainly acceptable in many applications, although it is somewhat swept under the rug here. On the technical side, the paper might have been strengthened by discussing in more detail why Julia is well-suited to this application.

We have added "steady-state" in multiple sentences to clarify (repeatedly) that this is the current scope of AIBECS. We have also added a sentence explaining why Julia was chosen.

> As for the github site and the code, both are of high quality. I worked through the 'age' tutorial, and it is straightforward and informative -- a model of how to do this sort of thing, in my opinion.
> ### Minor comments
>
> * Overall, the citations are overly heavy, in my opinion. I prefer writing that
>       cites only papers that a reader might reasonably want to consult, to learn
>       more about a particular point raised in the text. As an example, L103
>       consists entirely of citations, and all readers are told about them is that
>       they relate to parameters relating to the carbon cycle. This seems to be
>       insufficient guidance, at least for readers like me.  I am not objecting to
>       this style, because I know it to be common in some fields.

We appreciate this comment, which is similar to a comment by @zhenwu0728. The purpose of this list of references was not to point the reader to more detailed studies of particular points. Instead, it was intended to provide links to previous relevant modelling work that could have benefited from a framework like the one provided by AIBECS. However, we agree that trimming was appropriate here, and we have thus removed most of that paragraph although we have kept references to a few high-profile publications. In our opinion, the manuscript now gets straighter to the point.

> * I suggest removing "ideal" from the title. Let readers decide that.

The title now reads "A tool (...)" instead of "The ideal tool (...)".

> * Para 1 (L2:18) I think it might be useful to add a few sentences on why Julia
>       is a good choice for this project. Otherwise, the paper might only be read by
>       folks who are already familiar with the language.

We have added "Written in [Julia](https://julialang.org/) — chosen for of its combined expressive power and efficiency" a bit later in the text, after introducing the (MATLAB) AWESOME OCIM.

> * L27: This suggestion by Box about all models being wrong was essentially
>       a discussion point, not a scientific finding or a mathematical conclusion. It
>       might help to rewrite that sentence in a form like "Box (1979) has remarked
>       that ", to make things a bit clearer.  In some existential sense,
>       perhaps _everything_ is wrong.

We have revised the text as per the suggestion.

> * L28: Readers might take the authors to mean that parameters _only_ arise when
>       assumptions are made.  This is not entirely true. Another use of parameters
>       is to use a model to make predictions for a range of cases.  While I do not
>       object to the phrasing as it is, I might suggest `L28 s/give rise/can give rise/`. (Here I am using a sort of `sed` instruction to indicate a suggested
>       edit, although the `28` refers to a line number as printed, not in the file.
>       I'll use that notation throughout this review.)

We have revised the text accordingly.

> * L31: It's not clear that minimizing model-observation misfit always improves
>       the skill of a model, if the latter is (I think, reasonably) taken to
>       indicate prediction of a yet-unobserved state.  Overfitting can yield poor
>       predictions, sometimes.  Perhaps `L32 s/which also improves/which may also improve/` might help a bit, if the authors agree that there is a slight
>       chance of confusion here.

We have revised the text to say "may improve".

> * L73: the `Taburet_etal_2019` citation is broken.

This citation has been fixed.

## Response to reviewer 2 (@zhenwu0728) comments

> I've worked through "ideal age" and "PO4-POP" examples, and they are well documented and informative. I only have a few minor comments for the package and the paper.
>
> * The examples are great, but Binder is not working properly (tried a few times to launch Binder and it crashed while running the model, possibly because of the limit of memory). I suggest the authors may use Pluto.jl and [PlutoSliderServer.jl](https://github.com/JuliaPluto/PlutoSliderServer.jl) to host examples.

It is correct that Binder does not work "well" for these tutorials. This is because one has to load AIBECS (and its dependencies) and download circulations and data on a remote machine.

Translating them to Pluto is a great idea but, unfortunately, the latest version of Pluto (v0.17.2) is *not* compatible with AIBECS.jl. The current parameter interface of AIBECS.jl (which works with earlier Pluto versions, such as v0.15.0) relies on [FieldMetadata.jl](https://github.com/rafaqz/FieldMetadata.jl) macros that extend functions that are imported from FieldMetadata.jl (e.g., `units`, `bounds`). However, in Pluto v0.17.2, imported functions now cannot be extended in separate cells (see [issue 846](https://github.com/fonsp/Pluto.jl/issues/846#issuecomment-760886664)).

In response, we have added a note at the start of each tutorial and each how-to guide mentioning that binder might be slow.

Eventually, we might entirely rewrite the parameter internals and the user interface (without macros), but we believe that such a revision is out of the scope of @zhenwu0728's suggestion.

> * A better `show()` is needed for `OCIM2.load()`, `F_and_∇ₓF`, etc. Now it's hard to get any useful information from the output.

We have fixed the output issues of `OCIM2.load()` by not returning the often-unused He-flux fields. These can still be loaded if the keyword argument `FeFluxes=true` is provided. We have also fixed the output of `F_and_∇ₓF` by revising the interface slightly. Where users previously called

```julia
F, ∇ₓF = F_and_∇ₓF(args...)
```

which returned two functions (`F` is the tendencies of the tracers and `∇ₓF` its Jacobian w.r.t the tracers),
now users should call

```julia
F = AIBECSFunction(args...)
```

which returns only `F`. The Jacobian is still accessible (by calling `F.jac(args...)`, following SciML syntax) but is not exposed to the users who generally don't need to directly use it. The output of `AIBECSFunction(args...)` is now cleaner and in line with the SciML ecosystem.

> For paper.md:
>
> * L11-15: It might be good to provide one or two examples for both simple and complicated models.

We have added the Archer et al. (2000) 2-box model example for the simple case and the MITgcm for an example of a complicated model.

> * L16: It would be good to point out that AIBECS.jl only resolves steady states.

We have added "steady-state" here (and multiple other places following @dankelley's comments).

> * L31: It is worth mentioning that minimizing model-observation mismatches may cause overfitting.

We have followed @dankelley's suggestion here and revised the text to simply say "*may* improve".

> * L73: `Taburet_etal_OcnSci_2019` citation is broken.

This citation has been fixed.

> * L102-119: I think there are too many citations, only one or two for each kind of model is enough.

This list has been cut down. (See response to the similar comment made by @dankelley above.)
