### 2021-12-10 382b4dd
JOSS paper (#87)
* JOSS paper

* Fix tests (updated WOA API)

* Fix SciML API (no Val{:jac})

* Fix #88
