+++ docs/paper.bib
@@ -8,7 +8,7 @@
 }
 
 
-@article{Archer_etal:2000,
+@article{Archer_etal_GBC_2000,
   author  = {Archer, David E. and Eshel, Gidon and Winguth, Arne and Broecker, Wallace and Pierrehumbert, Ray and Tobis, Michael and Jacob, Robert},
   title   = {Atmospheric p{CO}\textsubscript{2} sensitivity to the biological pump in the ocean},
   journal = {Global Biogeochemical Cycles},
@@ -77,6 +77,45 @@
   isbn      = {978-0-12-438150-6},
   doi       = {10.1016/B978-0-12-438150-6.50018-2},
   author    = {G.E.P. Box}
+}
+
+@software{Campin_etal_2021,
+  author       = {Jean-Michel Campin and
+                  Patrick Heimbach and
+                  Martin Losch and
+                  Gael Forget and
+                  edhill3 and
+                  Alistair Adcroft and
+                  amolod and
+                  Dimitris Menemenlis and
+                  dfer22 and
+                  Chris Hill and
+                  Oliver Jahn and
+                  Jeff Scott and
+                  stephdut and
+                  Matt Mazloff and
+                  Baylor Fox-Kemper and
+                  antnguyen13 and
+                  Ed Doddridge and
+                  Ian Fenty and
+                  Michael Bates and
+                  AndrewEichmann-NOAA and
+                  Timothy Smith and
+                  Torge Martin and
+                  Jonathan Lauderdale and
+                  Ryan Abernathey and
+                  samarkhatiwala and
+                  hongandyan and
+                  Bruno Deremble and
+                  dngoldberg and
+                  Pascal Bourgault and
+                  raphael dussin},
+  title        = {MITgcm/MITgcm: checkpoint67z},
+  year         = 2021,
+  publisher    = {Zenodo},
+  version      = {checkpoint67z},
+  doi          = {10.5281/zenodo.1409237},
+  url          = {https://doi.org/10.5281/zenodo.1409237}
 }
 
 @article{Chamberlain_etal_OM_2019,
@@ -154,7 +193,7 @@
   bibsource     = {dblp computer science bibliography, https://dblp.org}
 }
 
-@article{DeVries_Deutsch_NatGeo_2014,
+@article{DeVries_Deutsch_NatGeosci_2014,
   doi       = {10.1038/ngeo2300},
   year      = {2014},
   month     = nov,
@@ -167,7 +206,7 @@
   journal   = {Nature Geoscience}
 }
 
-@article{DeVries_etal_GBC_2013,
+@article{DeVries_etal_Biogeosciences_2013,
   author  = {DeVries, Tim and Deutsch, C. and Rafter, P. A. and Primeau, F.},
   title   = {Marine denitrification rates determined from a global {3-D} inverse model},
   journal = {Biogeosciences},
@@ -192,7 +231,7 @@
   journal   = {Nature}
 }
 
-@article{DeVries_Holzer:2019,
+@article{DeVries_Holzer_JGRO_2019,
   author  = {DeVries, Tim and Holzer, Mark},
   title   = {Radiocarbon and Helium Isotope Constraints on Deep Ocean Ventilation and Mantle-\textsuperscript{3}{He} Sources},
   journal = {Journal of Geophysical Research: Oceans},
@@ -203,7 +242,7 @@
   year    = {2019}
 }
 
-@article{DeVries_Primeau:2011,
+@article{DeVries_Primeau_JPO_2011,
[…]
