+++ docs/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'AIBECS.jl: The ideal tool for exploring global marine biogeochemical cycles.'
+title: 'AIBECS.jl: A tool for exploring global marine biogeochemical cycles.'
 tags:
   - Julia
   - Oceanography
@@ -8,19 +8,21 @@ tags:
 authors:
   - name: Benoît Pasquier^[corresponding author]
     orcid: 0000-0002-3838-5976
-    affiliation: "1" # (Multiple affiliations must be quoted)
-  - name: Seth G. John
-    orcid: 0000-0002-8257-626X
-    affiliation: "1" # (Multiple affiliations must be quoted)
+    affiliation: "1,3" # (Multiple affiliations must be quoted)
   - name: François W. Primeau
     orcid: 0000-0001-7452-9415
     affiliation: "2"
+  - name: Seth G. John
+    orcid: 0000-0002-8257-626X
+    affiliation: "1" # (Multiple affiliations must be quoted)
 affiliations:
   - name: Department of Earth Sciences, University of Southern California
     index: 1
   - name: Department of Earth System Science, University of California, Irvine
     index: 2
-date: 19 Aug 2021
+  - name: (now at) School of Mathematics and Statistics, University of New South Wales, Sydney
+    index: 3
+date: 7 Dec 2021
 bibliography: paper.bib
 
 ---
@@ -33,9 +35,9 @@ bibliography: paper.bib
 The ocean transports, mixes, and transforms chemical constituents on a multitude of time and length scales.
 Observations and models are both essential for making sense of this complex system.
 However, the days of just publishing data without any quantitative modelling are over, increasing pressure on sea-going oceanographers that are expected to be proficient in the use of biogeochemical models.
-And while the simplest of these models consist of only a few boxes, with solutions that are easily obtained by hand or on a simple desktop computer, the most advanced models have high-resolution three-dimensional meshes that require high-performance computing (HPC) clusters and a considerable amount of computational-science expertise.
+And while the simplest of these models consist of only a few boxes [e.g., the two-box model of @Archer_etal_GBC_2000], with solutions that are easily obtained by hand or on a simple desktop computer, the most advanced models have high-resolution three-dimensional meshes that require high-performance computing (HPC) clusters and a considerable amount of computational-science expertise [e.g., the MITgcm, @Campin_etal_2021].
 A high barrier to entry for modelling in oceanography hinders advances in the field.
-[AIBECS.jl](https://github.com/JuliaOcean/AIBECS.jl) [@Pasquier:2021] is a [JuliaOcean](https://github.com/JuliaOcean/)-affiliated package written in [Julia](https://julialang.org/) [@Bezanson_etal:2017], which aims to lower this barrier by providing an easy-to-use, open-source, and modular framework for simulating global marine tracers and biogeochemical parameter fitting.
+[AIBECS.jl](https://github.com/JuliaOcean/AIBECS.jl) [@Pasquier:2021] is a [JuliaOcean](https://github.com/JuliaOcean/)-affiliated package written in [Julia](https://julialang.org/) [@Bezanson_etal:2017], which aims to lower this barrier by providing an easy-to-use, open-source, and modular framework for simulating global marine tracers in steady-state and biogeochemical parameter fitting.
 
 
 
@@ -45,21 +47,20 @@ A high barrier to entry for modelling in oceanography hinders advances in the fi
 A conceptual model of the cycle of any marine tracer essentially requires two components.
[…]
