### 2021-12-16 28e7a8a
CompatHelper: bump compat for Bijectors to 0.10, (keep existing compat) (#89)
Co-authored-by: CompatHelper Julia <compathelper_noreply@julialang.org>
