+++ Project.toml
@@ -36,7 +36,7 @@ Unitful = "1986cc42-f94f-5a68-af5c-568840ba703d"
 UnitfulRecipes = "42071c24-d89e-48dd-8a24-8a12d9b8861f"
 
 [compat]
-Bijectors = "0.8, 0.9"
+Bijectors = "0.8, 0.9, 0.10"
 CodecZlib = "0.7"
 DataDeps = "0.7"
 DataFrames = "0.20, 0.21, 0.22, 1.0"
