### 2021-12-16 #3814 @kthyng
@briochemc can you archive your software in a place like Zenodo and report back the doi here? Also please update the metadata in the archive so the author list and title exactly match your JOSS paper.
Also, please state here the software version you would like associated with your JOSS paper — you'd probably want to make a new tag/release of the software also.
