### 2021-12-16 #3814 @kthyng
@briochemc The paper looks good, but:
* is reference Garcia et al a bit repetitive?
* Kelley 2003 should have Newton capitalized
* Should julia be capitalized? (Rackaukas )
* Schlitzer - capitalize Atlantic and uncapitalize the journal
* Thacker - I think Hessian should be capitalized?
* please look through all references in detail to catch other errors
