### 2021-12-17 5154904
Tag new Zenodo release + minor fixes for JOSS
