+++ docs/paper.bib
@@ -1,4 +1,4 @@
-@article{Amante_Eakins_2009,
+@dataset{Amante_Eakins_2009,
   title     = {{ETOPO1} 1 arc-minute global relief model: procedures, data sources and analysis},
   author    = {Amante, Christopher and Eakins, Barry W},
   year      = {2009},
@@ -141,9 +141,9 @@
   year     = {2016}
 }
 
-@misc{Dai_2017,
+@dataset{Dai_2017,
   author    = {Aiguo {Dai}},
-  title     = {Dai and Trenberth Global River Flow and Continental Discharge Dataset},
+  title     = {Global River Flow and Continental Discharge Dataset},
   publisher = {Research Data Archive at the National Center for Atmospheric Research, Computational and Information Systems Laboratory},
   address   = {Boulder CO},
   year      = {2017},
@@ -290,7 +290,7 @@
   year    = {2014}
 }
 
-@misc{Distributions.jl-2019,
+@software{Distributions.jl-2019,
   author = {Dahua Lin and
                   John Myles White and
                   Simon Byrne and
@@ -532,7 +532,7 @@
 @inbook{Kelley_2003_1,
   author    = {Kelley, C. T.},
   chapter   = {1. Introduction},
-  title     = {Solving Nonlinear Equations with Newton's Method},
+  title     = {Solving Nonlinear Equations with {Newton}'s Method},
   pages     = {1-25},
   doi       = {10.1137/1.9780898718898.ch1},
   year      = {2003},
@@ -661,13 +661,12 @@
   author  = {Xingwen Li and François Primeau}
 }
 
-@misc{Luijendijk_etal_2019,
+@dataset{Luijendijk_etal_2019,
   author    = {Elco {Luijendijk} and Tom {Gleeson} and Nils {Moosdorf}},
   title     = {{Geospatial data and model results for a global model study of coastal groundwater discharge}},
   year      = {2019},
   doi       = {10.1594/PANGAEA.907641},
   note      = {Supplement to: Luijendijk, E et al. (2019): The Flow of Fresh Groundwater and Solutes to the World's Oceans and Coastal Ecosystems. EarthArXiv Preprints, https://doi.org/10.31223/osf.io/sw8r4},
-  type      = {data set},
   publisher = {PANGAEA}
 }
 
@@ -794,13 +793,15 @@ consistent data product for the world ocean},
 }
 
 @article{Rackauckas_Nie_JORS_2017,
-  title     = {Differentialequations.jl--a performant and feature-rich ecosystem for solving differential equations in julia},
+  title     = {{DifferentialEquations.jl --- A performant and feature-rich ecosystem for solving differential equations in Julia}},
   author    = {Rackauckas, Christopher and Nie, Qing},
   journal   = {Journal of Open Research Software},
   volume    = {5},
   number    = {1},
   year      = {2017},
-  publisher = {Ubiquity Press}
+  pages     = {15},
+  publisher = {Ubiquity Press},
+  doi       = {http://doi.org/10.5334/jors.151}
 }
 
 @article{RevelsLubinPapamarkou2016,
@@ -844,7 +845,7 @@ consistent data product for the world ocean},
 }
 
 
-@article{Schlitzer_etal_ChemGeo_2018,
+@dataset{Schlitzer_etal_ChemGeo_2018,
[…]
