### 2021-12-17 #3814 @briochemc
- [ ] archived in Zenodo [here](XXX): doi:XXX
- [x] author list and title now exactly match the JOSS paper
- [x] software version to be associated with the JOSS paper: v0.11.2
- [x] Garcia et al reference is fixed
- [x] Kelley 2003 reference fixed (Newton capitalized)
- [x] Rackaukas reference fixed (Julia capitalized)
- [x] Schlitzer reference fixed (Atlantic capitalized and journal name uncapitalized)
- [x] Thacker reference fixed (Hessian capitalized)
- [x] please look through all references in detail to catch other errors. Fixed the following (I might have fixed some references that do not make it into the paper as well):
  - [x] Amante and Eakins (2009) entry type fixed to be a `@dataset` entry type
  - [x] Dai (2017) fixed (now a `@dataset` and redundant name in title removed)
  - [x] Lin et al (2019) fixed (now a `@software`)
  - [x] Luijendijk (2019) fixed (now a `@dataset`)
  - [x] Rackaukas and Nie (2017) fixed (on top of editor request, DifferentialEquations.jl was not capitalized properly and page number and DOI were missing)
  - [x] Schlitzer et al (2018) fixed (now a `@dataset`)
